Data.Rope.UTF16

Documentation

data Rope

A SplayTree of Text values optimised for being indexed by and modified at UTF-16 code units and row/column (RowColumn) positions. Internal invariant: No empty Chunks in the SplayTree

Queries

null :: Rope -> Bool

Is the rope empty?

Since: 0.2.0.0

length :: Rope -> Int

Length in code units (not characters)

rows :: Rope -> Int

The number of newlines in the rope

Since: 0.3.0.0

columns :: Rope -> Int

The number of code units (not characters) since the last newline or the start of the rope

Since: 0.3.0.0

Conversions

toText :: Rope -> Text

toLazyText :: Rope -> Text

fromText :: Text -> Rope

toString :: Rope -> String

Transformations

map :: (Char -> Char) -> Rope -> Rope

Map over the characters of a rope

Since: 0.3.0.0

intercalate :: Rope -> [Rope] -> Rope

Concatenate the interspersion of a rope between the elements of a list of ropes

Since: 0.3.0.0

Chunking

toChunks :: Rope -> [Text]

The raw Text data that the Rope is built from

unconsChunk :: Rope -> Maybe (Text, Rope)

Get the first chunk and the rest of the Rope if non-empty

unsnocChunk :: Rope -> Maybe (Rope, Text)

Get the last chunk and the rest of the Rope if non-empty

UTF-16 code unit indexing

splitAt :: Int -> Rope -> (Rope, Rope)

Split the rope at the nth code unit (not character)

take :: Int -> Rope -> Rope

Take the first n code units (not characters)

drop :: Int -> Rope -> Rope

Drop the first n code units (not characters)

data RowColumn

Constructors

RowColumn

Fields

row :: !Int

Number of newlines before this position

column :: !Int

Number of code units since last newline or start of string

rowColumnCodeUnits :: RowColumn -> Rope -> Int

Get the code unit index in the rope that corresponds to a RowColumn position

Since: 0.2.0.0

codeUnitsRowColumn :: Int -> Rope -> RowColumn

Get the RowColumn position that corresponds to a code unit index in the rope

Since: 0.3.2.0

splitAtLine :: Int -> Rope -> (Rope, Rope)

Split the rope immediately after the i:th newline

Since: 0.3.1.0

Breaking by predicate

span :: (Char -> Bool) -> Rope -> (Rope, Rope)

span f r = (takeWhile f r, dropWhile f r)

break :: (Char -> Bool) -> Rope -> (Rope, Rope)

break f = span (not . f)

takeWhile :: (Char -> Bool) -> Rope -> Rope

takeWhile f = fst . span f

dropWhile :: (Char -> Bool) -> Rope -> Rope

dropWhile f = snd . span f

Folds

foldl :: (a -> Char -> a) -> a -> Rope -> a

Fold left

Since: 0.3.0.0

foldl' :: (a -> Char -> a) -> a -> Rope -> a

A strict version of foldl

Since: 0.3.0.0

foldr :: (Char -> a -> a) -> a -> Rope -> a

Fold right

Since: 0.3.0.0

Special folds

any :: (Char -> Bool) -> Rope -> Bool

Do any characters in the rope satisfy the predicate?

Since: 0.3.0.0

all :: (Char -> Bool) -> Rope -> Bool

Do all characters in the rope satisfy the predicate?

Since: 0.3.0.0